requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.5.0
ultralytics>=8.0.0

# Phiên bản mới nhất tương thích, tránh xung đột
langchain-core>=0.0.267
langchain>=0.0.267
langgraph>=0.4.5
langgraph-checkpoint>=0.0.1
langchain-openai>=0.0.2

fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.5
pydantic>=2.0.0
matplotlib>=3.4.0
scikit-image>=0.18.0
scikit-learn>=1.0.0
tqdm>=4.60.0
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.23.0
